fix!: Move `wxt/storage` to `wxt/utils/storage` (#1271)
