+++ packages/wxt/src/utils/storage.ts
@@ -1,5 +1,5 @@
 /**
  * Re-export the [`@wxt-dev/storage` package](https://www.npmjs.com/package/@wxt-dev/storage).
- * @module wxt/storage
+ * @module wxt/utils/storage
  */
 export * from '@wxt-dev/storage';
